mapilm_backend/requirements.txt
Django==5.0.4
dj-database-url==2.1.0
djangorestframework==3.15.1
django-channels==4.1.0
channels-redis==4.2.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-decouple==3.8
firebase-admin==6.5.0
Pillow==10.3.0
celery==5.4.0
redis==5.0.4
drf-spectacular==0.27.2
gunicorn==22.0.0
daphne==4.1.2
whitenoise==6.6.0
cloudinary==1.40.0
django-cloudinary-storage==0.3.0
cryptography==42.0.7
